+++ src/java/com/jaeksoft/searchlib/crawler/database/url/UrlManagerJdbc.java
@@ -362,7 +362,7 @@ public class UrlManagerJdbc extends UrlManager {
 	@Override
 	public void getUrls(String like, String host, FetchStatus fetchStatus,
 			ParserStatus parserStatus, IndexStatus indexStatus, Date startDate,
-			Date endDate, Field orderBy, int start, int rows, UrlList urlList)
+			Date endDate, Field orderBy, long start, long rows, UrlList urlList)
 			throws CrawlDatabaseException {
 		Transaction transaction = null;
 		try {
@@ -370,8 +370,8 @@ public class UrlManagerJdbc extends UrlManager {
 			Query query = getUrls(transaction, like, host, fetchStatus,
 					parserStatus, indexStatus, startDate, endDate, orderBy,
 					true);
-			query.setFirstResult(start);
-			query.setMaxResults(rows);
+			query.setFirstResult((int) start);
+			query.setMaxResults((int) rows);
 			List<UrlItem> results = (List<UrlItem>) query
 					.getResultList(UrlItem.class);
 			urlList.getPartialList().set(results, query.getResultCount());
